+++ server.js
@@ -8,11 +8,16 @@ const SequelizeStore = require('connect-session-sequelize')(session.Store);
 require('dotenv').config();
 const { view_routes, auth } = require('./routes/index')
 const app = express();
+// const quoteFunction = require('axios_routes');
+const makeGetRequest = require('./routes/axios_routes');
 app.use(express.static(path.join('front')));
 app.engine('hbs', engine({ extname: '.hbs'}));
 app.set('view engine', 'hbs');
 app.use(express.json());
 app.use(express.urlencoded({ extended: false }));
+
+
+makeGetRequest();
 
 app.use(session({
     secret: process.env.SESSION_SECRET,
@@ -31,6 +36,7 @@ app.get('/', (req, res) => {
     res.render ("index")
 })
 
-db.sync({ force: false }).then(() => {
+db.sync({ force: true }).then(() => {
     app.listen(PORT, () => console.log(`Listening on Port ${PORT}`));
     });
+
