+++ views/index.hbs
@@ -89,7 +89,7 @@
         Inspiration
       </div>
       <div class="card-body">
-        <div class="list-group list-group-flush">Quote goes here</div>
+        <div class="list-group list-group-flush" id="placement">Quote goes here</div>
       </div>
       <div class="card-footer">
         <button id="genBtn">Motivate Me!</button>
